demo/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
earthengine-api
requests
Pillow
numpy
fastmcp
qwen-agent
